Cleaned the waitlist States function
